Run MySQL setup only once

+++ setup.sh
@@ -26,10 +26,14 @@ then
     done
 fi
 
+if [ ! -d mysql/swgemu ]
+then
 echo "Setting up MySQL Database"
-mysql_container=$(docker run -d --rm -e MYSQL_ALLOW_EMPTY_PASSWORD=true -v $PWD/mysql:/var/lib/mysql -v $PWD:/work mysql)
-sleep 30
-docker exec -it $mysql_container sh -c "/work/mysql-setup.sh"
-docker stop $mysql_container > /dev/null
+    mysql_container=$(docker run -d --rm -e MYSQL_ALLOW_EMPTY_PASSWORD=true -v $PWD/mysql:/var/lib/mysql -v $PWD:/work mysql)
+    sleep 30
+    docker exec -it $mysql_container sh -c "/work/mysql-setup.sh"
+    docker stop $mysql_container > /dev/null
+fi
 
-echo "Run ./engine3-build.sh and ./build.sh"
+echo "Done!"
+echo "Run ./engine3-build.sh to build engine and run ./build.sh Core3"
